devtoolbox/pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "devtoolbox"
version = "0.1.0"
description = "Dev Toolbox CLI with small daily dev utilities."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]

dependencies = [
  "typer>=0.9.0",
  "markdown>=3.6",
  "weasyprint>=61.0",
  "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.4.0",
]

[project.scripts]
dt = "devtoolbox.cli:app"

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
